Get list of authenticated sessions from the session table
In Trac 1.0 the "session_attributes" table contained the sid (i.e. username) of
each authenticated user. This does not seem to be the case anymore with Trac
1.2: only the "session" table gets updated when the user logins the first time.
The "session_attribute" table is only(?) updated when user modifies his/her
preferences (e.g. adds "Full name" or "Email address"). With this commit the
list of users who have logged in is taken from the "session" table, so that
emails in "session_attributes" table can updated.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>